requirements.txt
asgiref==3.5.2
Django==4.1.3
django-taggit==3.0.0
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
